requirements.txt
python-dateutil==2.7.3
requests==2.8.1
PyYAML==3.11
argparse==1.2.1
batcher==2
boto==2.39.0
dateparser==0.3.0
gglsbl==0.3
jdatetime==1.7.1
mock==1.3.0
momoko==2.2.2
moto==0.4.21
nose2==0.5.0
numpy==1.14.5
pandas==0.23.1
scipy==1.1.0
scikit-learn==0.19.1
sklearn-pandas==1.6.0
six==1.10.0
tornado==4.2.1
unicodecsv==0.14.1
testfixtures==4.9.0
tldextract==2.0.1
raven==5.23.0
matplotlib==2.1.1
eli5==0.8
#-e git+https://anomali-build-read:ccfd59e1464993067c097628d07f02d7dabab5e6@ghe.anomali.com/ThreatStream/domaingeneration.git@python2-master#egg=domaingeneration
protobuf==3.7.1
nolds==0.5.1
treeinterpreter==0.2.2
boto3==1.9.182
APScheduler==3.6.0
PyGreSQL==4.2
keras==2.2.4
tensorflow==1.13.2